requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0
pydantic==2.5.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyPDF2==3.0.1
ebooklib==0.18
python-docx==1.1.0
aiofiles==23.2.1
jinja2==3.1.2
openai==1.3.0
httpx==0.25.2
websockets==12.0
python-dotenv==1.0.0